Add files via upload
This is the best version of our website for the front end user. This code contains code in HTML and Cascading style sheet with some JavaScript functioning. If any problem occurs in this version please let us know or commit! This is just a prototype of a functional website! You can just pull request for any changes! And path of file should be get changed!